Refactor attendance types and integrate attendance router into posting routes
