+++ server/src/api/routes/organization/attendance.types.ts
@@ -1,5 +1,13 @@
-export type {
-  OrganizationPostingAttendanceBulkUpdateResponse,
-  OrganizationPostingAttendanceResponse,
-  OrganizationPostingEnrollmentAttendanceUpdateResponse,
-} from './posting.types.js';
+import { OrganizationPostingWithoutVectors } from '../../../db/tables.js';
+import { PostingEnrollment, SuccessResponse } from '../../../types.js';
+
+export type OrganizationPostingEnrollmentAttendanceUpdateResponse = SuccessResponse;
+
+export type OrganizationPostingAttendanceResponse = {
+  posting: Pick<OrganizationPostingWithoutVectors, 'id' | 'title' | 'location_name'>;
+  enrollments: PostingEnrollment[];
+};
+
+export type OrganizationPostingAttendanceBulkUpdateResponse = {
+  updated_count: number;
+};
