+++ server/src/api/routes/organization/index.ts
@@ -1,7 +1,6 @@
 import { Router, Response } from 'express';
 import zod from 'zod';
 
-import attendanceRouter from './attendance.js';
 import {
   OrganizationCrisisResponse,
   OrganizationCrisesResponse,
@@ -277,6 +276,5 @@ organizationRouter.put('/profile', async (req, res) => {
 organizationRouter.post('/reset-password', resetPassword);
 
 organizationRouter.use('/posting', postingRouter);
-organizationRouter.use('/posting', attendanceRouter);
 
 export default organizationRouter;
